ENH Moved TestOnly classes to tests dir
- Replace DataExtension with Extension
- Used create instead of fixture entry in test
- Simplified engine extension check
