+++ src/Extension/Engine/SiteTreePublishingEngine.php
@@ -279,10 +279,7 @@ class SiteTreePublishingEngine extends SiteTreeExtension implements Resettable
             }
 
             // The page does not include the required extension, and it doesn't implement a Trigger
-            if (!$siteTree->hasExtension(PublishableSiteTree::class)
-                && !$siteTree->hasExtension(StaticPublishingTrigger::class)
-                && !($siteTree instanceof StaticPublishingTrigger)
-            ) {
+            if (!$siteTree->hasExtension(StaticPublishingTrigger::class) && !($siteTree instanceof StaticPublishingTrigger)) {
                 return;
             }
